fixed renaming typo
